Add context menu to insert menu
