+++ src/frontend/cards/insertable-card.tsx
@@ -15,6 +15,7 @@ import {
     InsertableOut,
     LibraryOut
 } from "../../shared/api-models";
+import { Configuration } from "../../shared/configuration-models";
 import { ElementType } from "../../shared/types";
 import { SearchHit } from "../search/search";
 import {
@@ -112,26 +113,35 @@ export function InsertableCard(props: InsertableCardProps): ReactNode {
 interface InsertableMenuItemsProps {
     favorite: Favorite | undefined;
     insertable: InsertableOut;
+    inInsertMenu?: boolean;
+    configuration?: Configuration;
 }
 
 export function InsertableMenuItems(
     props: InsertableMenuItemsProps
 ): ReactNode {
-    const { favorite, insertable } = props;
+    const { favorite, insertable, inInsertMenu, configuration } = props;
 
     return (
         <>
-            <QuickInsertItems
-                insertable={insertable}
-                isFavorite={favorite !== undefined}
-            />
-            <Menu.Divider />
+            {!inInsertMenu && (
+                <>
+                    <QuickInsertItems
+                        insertable={insertable}
+                        isFavorite={favorite !== undefined}
+                    />
+                    <Menu.Divider />
+                </>
+            )}
             <FavoriteInsertableItem
                 favorite={favorite}
                 insertable={insertable}
             />
             <Menu.Divider />
-            <OpenDocumentItems path={insertable.path} />
+            <OpenDocumentItems
+                path={insertable.path}
+                configuration={configuration}
+            />
             <AdminOptionsSubmenu>
                 <InsertableAdminContextMenu insertable={insertable} />
             </AdminOptionsSubmenu>
